Bump cosmic-text 0.18 -> 0.19, wgpu 28 -> 29
wgpu 29 API changes:
- PipelineLayoutDescriptor.bind_group_layouts now &[Option<&_>]
- Instance::new takes InstanceDescriptor by value
- InstanceDescriptor::default removed; use new_without_display_handle
- Surface::get_current_texture returns CurrentSurfaceTexture enum
- BufferViewMut no longer derefs to &mut [u8]; use .slice(..len)
- max_storage_buffer_binding_size widened to u64

cosmic-text 0.19 lazy setters drop the font_system arg from
Buffer::set_text, set_rich_text, and set_size.

Also bumps wgpu-profiler 0.26 -> 0.27 to match wgpu 29.

Coded by an LLM.

+++ src/text_atlas.rs
@@ -252,7 +252,7 @@ impl TextAtlas {
         let new_end = glyph_offset + prepared.blob_size;
         if new_end > self.buffer_capacity {
             let required_bytes = new_end as u64 * BYTES_PER_TEXEL;
-            let max_bytes = device.limits().max_storage_buffer_binding_size as u64;
+            let max_bytes = device.limits().max_storage_buffer_binding_size;
             if required_bytes > max_bytes {
                 return Err(crate::types::PrepareError::AtlasFull);
             }
